Move conversions and legacy marshalling to the legacy package
This also wraps up the code for converting the legacy genesis block
